+++ js/main.js
@@ -1,3 +1,1 @@
-import { createPhotos, PHOTO_COUNT } from './create-photos';
-
-createPhotos(PHOTO_COUNT);
+import './gallery.js';
